Fix gradle config cache
